ci: Add arm64 flatpak build.

+++ ci/circleci-build-flatpak.sh
@@ -9,34 +9,57 @@ MANIFEST=$(cd flatpak; ls org.opencpn.OpenCPN.Plugin*yaml)
 echo "Using manifest file: $MANIFEST"
 set -x
 
-# Give the apt update daemons a chance to leave the scene while we build.
-sudo systemctl stop apt-daily.service apt-daily.timer
-sudo systemctl kill --kill-who=all apt-daily.service apt-daily-upgrade.service
-sudo systemctl mask apt-daily.service apt-daily-upgrade.service
-sudo systemctl daemon-reload
+# On old systems: give the apt update daemons a chance to leave the scene
+# while we build.
+if systemctl status apt-daily-upgrade.service &> /dev/null; then
+    sudo systemctl stop apt-daily.service apt-daily.timer
+    sudo systemctl kill --kill-who=all \
+        apt-daily.service apt-daily-upgrade.service
+    sudo systemctl mask apt-daily.service apt-daily-upgrade.service
+    sudo systemctl daemon-reload
+fi
 
 # Install the flatpak PPA so we can access a usable flatpak version
-# (not available on Ubuntu Xenial 16.04, the VM this runs on)
-sudo add-apt-repository -y ppa:alexlarsson/flatpak
-wget -q -O - https://dl.google.com/linux/linux_signing_key.pub \
-    | sudo apt-key add -
+# (not available on Ubuntu Xenial 16.04)
+if [ -n "$USE_CUSTOM_PPA" ]; then
+    sudo add-apt-repository -y ppa:alexlarsson/flatpak
+    wget -q -O - https://dl.google.com/linux/linux_signing_key.pub \
+        | sudo apt-key add -
+fi
 sudo apt update
 
-# Install the dependencies: openpcn and the flatpak SDK.
-sudo apt install build-essential flatpak-builder flatpak tar
-flatpak remote-add --user --if-not-exists flathub \
-    https://flathub.org/repo/flathub.flatpakrepo
-flatpak install --user -y flathub org.opencpn.OpenCPN > /dev/null
-flatpak install --user -y flathub org.freedesktop.Sdk//18.08  >/dev/null
+sudo apt install flatpak flatpak-builder
+flatpak remote-add --user --if-not-exists \
+    flathub https://dl.flathub.org/repo/flathub.flatpakrepo
 
-# Patch the runtime version so it matches the nightly builds'
-sed -i '/^runtime-version/s/:.*/: stable/' flatpak/$MANIFEST
+
+# For now, horrible hack: aarch 64 builds are using the updated runtime
+# 20.08 and the opencpn beta version using old 18.08 runtime.
+if [ "$FLATPAK_BRANCH" = 'beta' ]; then
+        flatpak install --user -y flathub org.freedesktop.Sdk//20.08 >/dev/null
+        flatpak remote-add --user --if-not-exists flathub-beta \
+            https://flathub.org/beta-repo/flathub-beta.flatpakrepo
+        flatpak install --user -y --or-update flathub-beta \
+            org.opencpn.OpenCPN >/dev/null
+        sed -i '/sdk:/s/18.08/20.08/'  flatpak/org.opencpn.*.yaml
+else
+        flatpak install --user -y flathub org.freedesktop.Sdk//18.08 >/dev/null
+        flatpak remote-add --user --if-not-exists flathub \
+            https://flathub.org/repo/flathub.flatpakrepo
+        flatpak install --user -y --or-update flathub \
+            org.opencpn.OpenCPN >/dev/null
+        FLATPAK_BRANCH='stable'
+fi
+
+# Patch the runtime version so it matches the nightly builds
+# or beta as appropriate.
+sed -i "/^runtime-version/s/:.*/: $FLATPAK_BRANCH/" flatpak/$MANIFEST
 
 # The flatpak checksumming needs python3:
 pyenv local $(pyenv versions | sed 's/*//' | awk '{print $1}' | tail -1)
 cp .python-version $HOME
 
-# Install a recent python, flatpak's is too old.
+# Install a recent cmake, ubuntu 16.04 is too old.
 export PATH=$HOME/.local/bin:$PATH
 python -m pip install --user cmake
 
@@ -53,7 +76,7 @@ git checkout ../flatpak/$MANIFEST
 echo -n "Waiting for apt_daily lock..."
 sudo flock /var/lib/apt/daily_lock echo done
 
-# Install cloudsmith, requiered by upload script
+# Install cloudsmith, required by upload script
 python3 -m pip install --user --upgrade pip
 python3 -m pip install --user cloudsmith-cli
